update add to cart in menu
